tests: support unix sockets for PGHOST
