+++ test/legacy/conftest.py
@@ -1,29 +1,8 @@
 import sys
-from os import environ
 
 import pg8000
 
 import pytest
-
-
-@pytest.fixture(scope="class")
-def db_kwargs():
-    db_connect = {
-        'user': 'postgres',
-        'password': 'pw'
-    }
-
-    for kw, var, f in [
-                ('host', 'PGHOST', str),
-                ('password', 'PGPASSWORD', str),
-                ('port', 'PGPORT', int)
-            ]:
-        try:
-            db_connect[kw] = f(environ[var])
-        except KeyError:
-            pass
-
-    return db_connect
 
 
 @pytest.fixture
